Use a qp-trie for domain name lookups
This change allows you to configure NSD to use a DNS-specific qp-trie
instead of either a radtree or rbtree for domain name and zone
lookups.

The qp-trie provides slightly better performance than the radtree with
much lower memory usage, a little more than the rbtree.

This variety of qp-trie gets some of its compactness by using node
sizes tuned to fit the common hostname character set, and an escaping
scheme that prepares DNS names for lookup in a way that simplifies the
data structure. The qp-trie implementation is less than half the code
of the radtree (not counting the explanatory essay in qp-trie.h).

For best performance, this code should be optimised for a target CPU
architecture that supports the native popcount instruction.