test and benchmark code for compact lookup keys
tpkg/fuzz has a libfuzzer harness for testing radname_d2r() and
radname_r2d(). It was written to verify compact lookup keys, but the
harness works for the normal versions too. I'm afraid the Makefile
support is horrible, the bare minimum to get it working.

The treeperf benchmark doesn't need many changes for compact keys, but
it needed some robustification. It is now more careful to avoid
problematic domain names in the input file. It also uses a faster
random number generator because random_generate() turned out to be
slow enough on a VM to dominate the benchmark timings - that is,
benchmark runs on this machine take about 0.4s - 1.9s and
random_generate() increased that by about 2 seconds.

I previously had an ugly hack to silence a warning in -flto builds,
but it made normal builds fail, so I ditched the hack. The warning
points out that the treeperf code is making some assumptions about how
linking works, and I think those assumptions are safe, but the warning
is worth keeping.