a big refactor
Add a layer of indirection between the owner of the tree, so that we
can swap the tree with a simple atomic store.

Make the tree root and allocator bumf private.

Improve the names of some types and variables.

Separate compact() and release() because compact() needs to happen
before the swap and release() after the swap.

Add a `keep` counter for each page, so that when there is a large
transaction we can do intermediate GCs without interfering with the
live read-only pages.